Add proper detection for all Opal models and release 0.4.0
There are 32 different Opal models according to the webOS Doctors I was able to get my hands on. Add a similar table as for the Topaz models.
The main differnce is that both the WiFi and 3G/4G models of the Opals will require the same key handling as the 3G Topaz models (due to the usage of the Mini PCI Express cards). Therefore we set all Opal boards to "3G".

Signed-off-by: Herman van Hazendonk <github.com@herrie.org>